icat4

+++ Code/TopCATCore/src/main/java/uk/ac/stfc/topcat/icatclient/v331/ICATInterfacev331.java
@@ -346,8 +346,8 @@ public class ICATInterfacev331 extends ICATWebInterfaceBase {
         if (datafile.getDatafileCreateTime() != null) {
             createDate = datafile.getDatafileCreateTime().toGregorianCalendar().getTime();
         }
-        return new TDatafile(serverName, datafile.getId().toString(), datafile.getName(), datafile.getFileSize(),
-                format, formatVersion, formatType, createDate, datafile.getLocation());
+        return new TDatafile(serverName, datafile.getId().toString(), datafile.getName(), datafile.getFileSize()
+                .longValue(), format, formatVersion, formatType, createDate, datafile.getLocation());
     }
 
 }
